Dashboard Project sidebar changes
