+++ apps/dashboard/src/app/(app)/team/[team_slug]/[project_slug]/(sidebar)/connect/in-app-wallets/_components/header.tsx
@@ -4,7 +4,7 @@ export async function InAppWalletsHeader() {
   return (
     <div className="">
       <h1 className="mb-2 font-semibold text-2xl tracking-tighter lg:text-3xl">
-        In-App Wallets
+        Wallets
       </h1>
       <p className="max-w-2xl text-muted-foreground text-sm leading-relaxed">
         A wallet infrastructure that enables apps to create, manage, and control
