--- apps/dashboard/src/app/(app)/team/[team_slug]/[project_slug]/(sidebar)/connect/in-app-wallets/_constants.ts
@@ -1,1 +1,0 @@
-export const TRACKING_CATEGORY = "team/in-app-wallets";
