+++ apps/dashboard/src/app/(app)/team/[team_slug]/[project_slug]/(sidebar)/engine/cloud/analytics/tx-chart/tx-chart-ui.tsx
@@ -185,7 +185,7 @@ function EmptyChartContent(props: {
               variant="primary"
               onClick={() => {
                 router.push(
-                  `/team/${props.teamSlug}/${props.project.slug}/engine/cloud/vault`,
+                  `/team/${props.teamSlug}/${props.project.slug}/vault`,
                 );
               }}
             >
