--- apps/dashboard/src/app/(app)/team/[team_slug]/[project_slug]/(sidebar)/nebula/components/analytics/utils.ts
@@ -1,13 +1,0 @@
-import { getNebulaFiltersFromSearchParams } from "lib/time";
-
-export function getNebulaAnalyticsRangeFromSearchParams(searchParams: {
-  from: string | undefined | string[];
-  to: string | undefined | string[];
-  interval: string | undefined | string[];
-}) {
-  return getNebulaFiltersFromSearchParams({
-    from: searchParams.from,
-    to: searchParams.to,
-    interval: searchParams.interval,
-  });
-}
