+++ apps/dashboard/src/app/(app)/team/[team_slug]/[project_slug]/(sidebar)/vault/components/list-access-tokens.client.tsx
@@ -20,7 +20,7 @@ import {
   SERVER_WALLET_MANAGEMENT_ACCESS_TOKEN_PURPOSE,
   createWalletAccessToken,
   initVaultClient,
-} from "../../lib/vault.client";
+} from "../../engine/cloud/lib/vault.client";
 
 export default function ListAccessTokens(props: {
   project: Project;
