+++ apps/dashboard/src/app/(app)/team/[team_slug]/[project_slug]/(sidebar)/vault/page.tsx
@@ -1,6 +1,6 @@
 import { getProject } from "@/api/projects";
+import { getAuthToken } from "@app/api/lib/getAuthToken";
 import { notFound } from "next/navigation";
-import { getAuthToken } from "../../../../../../../api/lib/getAuthToken";
 import { KeyManagement } from "./components/key-management";
 
 export default async function VaultPage(props: {
